+++ hapi-fhir-storage-batch2-jobs/src/main/java/ca/uhn/fhir/batch2/jobs/services/Batch2JobRunnerImpl.java
@@ -22,10 +22,10 @@ package ca.uhn.fhir.batch2.jobs.services;
 
 import ca.uhn.fhir.batch2.api.IJobCoordinator;
 import ca.uhn.fhir.batch2.api.JobOperationResultJson;
+import ca.uhn.fhir.batch2.jobs.export.BulkExportUtil;
 import ca.uhn.fhir.batch2.jobs.export.models.BulkExportJobParameters;
 import ca.uhn.fhir.batch2.model.JobInstance;
 import ca.uhn.fhir.batch2.model.JobInstanceStartRequest;
-import ca.uhn.fhir.batch2.model.StatusEnum;
 import ca.uhn.fhir.i18n.Msg;
 import ca.uhn.fhir.jpa.api.model.Batch2JobInfo;
 import ca.uhn.fhir.jpa.api.model.Batch2JobOperationResult;
@@ -33,7 +33,6 @@ import ca.uhn.fhir.jpa.api.model.BulkExportParameters;
 import ca.uhn.fhir.jpa.api.svc.IBatch2JobRunner;
 import ca.uhn.fhir.jpa.batch.models.Batch2BaseJobParameters;
 import ca.uhn.fhir.jpa.batch.models.Batch2JobStartResponse;
-import ca.uhn.fhir.jpa.bulk.export.model.BulkExportJobStatusEnum;
 import ca.uhn.fhir.rest.server.exceptions.ResourceNotFoundException;
 import ca.uhn.fhir.util.Batch2JobDefinitionConstants;
 import org.slf4j.Logger;
@@ -97,31 +96,16 @@ public class Batch2JobRunnerImpl implements IBatch2JobRunner {
 	private Batch2JobInfo fromJobInstanceToBatch2JobInfo(@Nonnull JobInstance theInstance) {
 		Batch2JobInfo info = new Batch2JobInfo();
 		info.setJobId(theInstance.getInstanceId());
-		info.setStatus(fromBatchStatus(theInstance.getStatus()));
+		// we should convert this to a more generic enum (for all batch2)
+		// or use the status enum only (and not the bulk export enum)
+		// on the Batch2JobInfo
+		info.setStatus(BulkExportUtil.fromBatchStatus(theInstance.getStatus()));
 		info.setCancelled(theInstance.isCancelled());
 		info.setStartTime(theInstance.getStartTime());
 		info.setEndTime(theInstance.getEndTime());
 		info.setReport(theInstance.getReport());
 		info.setErrorMsg(theInstance.getErrorMessage());
 		return info;
-	}
-
-	public static BulkExportJobStatusEnum fromBatchStatus(StatusEnum status) {
-		switch (status) {
-			case QUEUED:
-			case FINALIZE:
-				return BulkExportJobStatusEnum.SUBMITTED;
-			case COMPLETED :
-				return BulkExportJobStatusEnum.COMPLETE;
-			case IN_PROGRESS:
-				return BulkExportJobStatusEnum.BUILDING;
-			default:
-				ourLog.warn("Unrecognized status {}; treating as FAILED/CANCELLED/ERRORED", status.name());
-			case FAILED:
-			case CANCELLED:
-			case ERRORED:
-				return BulkExportJobStatusEnum.ERROR;
-		}
 	}
 
 	private Batch2JobStartResponse startBatch2BulkExportJob(BulkExportParameters theParameters) {
